refactor(frontend): create components folder
